// Codes/cagmss-web-a4/src/app/pages/models/definitions.ts
export class Station {
    v01000: number;
    cStatName: string;
    cTypestation: string;
    cAera: string;
    cPrvoName: string;
    cCity: string;
    cCounty: string;
    cCode: string;
}
export class MangStatSet {
    v01000: number;
    v0100001: number;
    cBuscode: string;
    cRemark: string;
}
export class Definition {
    name: string;
    index: Array<number>;
}